+++ firebase_rules/lib/src/firebase/namespace/model/src/iterable.dart
@@ -1,0 +1,2 @@
+/// Base class for RulesSet and RulesList
+abstract class RulesIterable<T> {}